% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_errors.R
\name{list_errors}
\alias{list_errors}
\title{Get list of errors from Google Error Reporting}
\usage{
list_errors(project_id, api_key)
}
\arguments{
\item{project_id}{the project ID of your project on GCP}

\item{api_key}{an API key with permissions to write to Google Error Reporting}
}
\value{
No return, we focus on side effect
}
\description{
Get list of errors from Google Error Reporting
}
