% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_error.R
\name{report_error}
\alias{report_error}
\title{Report error to Google Error Reporting}
\usage{
report_error(message)
}
\arguments{
\item{message}{the error report to be written out to the}
}
\value{
No return, we focus on side effect
}
\description{
Report error to Google Error Reporting
}
